Move to very dynamic kryon stuff
The folder structure and variables are now read from a real folder structure.
This is probably not very secure, but it makes for easy setup of demo's if we wanted to add more folders and or automations.
The variables in the automation are now read from the file that represents the automation.
Every line is a variable name, there's probably an easy way to break this too. BUt hey, who cares!