Start using refresh token
We start using a refresh token and storing it as such. This allows us to have an `accesstoken.json` that we use in all the scripts, and also run `GetAuthCode.py` to work with the refresh token and update `accesstoken.json`
This moves the authentication stuff to a python script as it's becoming more advanced.

Receivetoken.py is updated to write both refresh_token and access token.